Added GenericParser to set the stage for other parsers
